Update
Because there was an using name space std in the main file and the .h files all used the "STD::" prefix instead it was confusing the system. I have updated the .cpp file to make it build with out these errors. Removing the using name space from the .cpp file.  and replacing it with std:: on all needed variables